emit location as state

+++ app/src/main/java/me/tbandawa/android/openweather/service/Coordinates.kt
@@ -1,6 +1,6 @@
 package me.tbandawa.android.openweather.service
 
 data class Coordinates(
-    var latitude: Long,
-    var longitude: Long
+    var latitude: Float,
+    var longitude: Float
 )
